web/flask/requirements.txt

flask==3.1.0
flask-cors==5.0.1


torch==2.5.1
torch-geometric==2.6.1
torchvision==0.20.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.3
matplotlib==3.9.4
h5py==3.12.1
networkx==3.2.1
sympy==1.13.1
joblib==1.4.2


requests==2.32.3
Pillow==11.2.1
PyMySQL==1.1.1
tqdm==4.67.1
triton==3.1.0
pefile==2024.8.26
requests-toolbelt==1.0.0
wordfreq
tld
xgboost